core/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.4